VisibilityType stores Int instead of Byte
Replaced NUMBER_OF_VISIBILITY_TYPES with VISIBILITY_TYPE_RANGE for better validation

+++ src/main/kotlin/io/fdeitylink/kero/file/ReadPxPack.kt
@@ -70,7 +70,7 @@ private fun Head.Companion.fromChannel(chan: ReadableByteChannel): Head {
                 val (visibilityType, scrollType) = ByteBuffer.allocate(2).let {
                     chan.read(it)
                     it.flip()
-                    Pair(VisibilityType(it.get()), ScrollType.values()[it.get().toUInt()])
+                    Pair(VisibilityType(it.get().toUInt()), ScrollType.values()[it.get().toUInt()])
                 }
 
                 Pair(it, LayerProperties(tileset, visibilityType, scrollType))
